Changed name of Stocks folder to stocks_detail
