+++ business_banking/lib/features/stocks_detail/api/stocks_service_response_model.dart
