+++ business_banking/lib/features/stocks_detail/bloc/stocks_bloc.dart
@@ -1,6 +1,6 @@
-import 'package:business_banking/features/stocks/api/stocks_service.dart';
-import 'package:business_banking/features/stocks/bloc/stocks_usecase.dart';
-import 'package:business_banking/features/stocks/model/stocks_view_model.dart';
+import 'package:business_banking/features/stocks_detail/api/stocks_service.dart';
+import 'package:business_banking/features/stocks_detail/bloc/stocks_usecase.dart';
+import 'package:business_banking/features/stocks_detail/model/stocks_view_model.dart';
 import 'package:clean_framework/clean_framework.dart';
 
 class StocksBloc extends Bloc {
@@ -15,6 +15,6 @@ class StocksBloc extends Bloc {
   StocksBloc({StocksService? stocksService}) {
     _useCase = StocksUseCase(
         (viewModel) => stocksViewModelPipe.send(viewModel as StocksViewModel));
-        stocksViewModelPipe.whenListenedDo(() => _useCase.create());
+    stocksViewModelPipe.whenListenedDo(() => _useCase.create());
   }
 }
