+++ business_banking/lib/features/stocks_detail/model/stocks_entity.dart
@@ -1,9 +1,8 @@
-import 'package:business_banking/features/stocks/api/stocks_service.dart';
+import 'package:business_banking/features/stocks_detail/api/stocks_service.dart';
 import 'package:clean_framework/clean_framework.dart';
 
 class StocksEntity extends Entity {
   final List<Map<String, dynamic>> stocks;
-  
 
   StocksEntity(
       {List<EntityFailure> errors = const [],
