+++ business_banking/lib/features/stocks_detail/model/stocks_view_model.dart
