+++ business_banking/lib/features/stocks_detail/ui/stocks_screen.dart
@@ -1,4 +1,4 @@
-import 'package:business_banking/features/stocks/model/stocks_view_model.dart';
+import 'package:business_banking/features/stocks_detail/model/stocks_view_model.dart';
 import 'package:clean_framework/clean_framework.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/framework.dart';
@@ -7,11 +7,13 @@ class StocksScreen extends Screen {
   final StocksViewModel viewModel;
 
   StocksScreen({required this.viewModel});
-  
 
   @override
   Widget build(BuildContext context) {
-    return Container(child: Center(child: Text('Stocks'),),);
+    return Container(
+      child: Center(
+        child: Text('Stocks'),
+      ),
+    );
   }
-  
 }
