+++ business_banking/test/features/stocks/api/stocks_service_adapter_test.dart
@@ -1,7 +1,6 @@
-
-import 'package:business_banking/features/stocks/api/stocks_service_response_model.dart';
-import 'package:business_banking/features/stocks/bloc/stocks_service_adapter.dart';
-import 'package:business_banking/features/stocks/model/stocks_entity.dart';
+import 'package:business_banking/features/stocks_detail/api/stocks_service_response_model.dart';
+import 'package:business_banking/features/stocks_detail/bloc/stocks_service_adapter.dart';
+import 'package:business_banking/features/stocks_detail/model/stocks_entity.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
