+++ business_banking/test/features/stocks/api/stocks_service_test.dart
@@ -1,5 +1,5 @@
-import 'package:business_banking/features/stocks/api/stocks_service.dart';
-import 'package:business_banking/features/stocks/api/stocks_service_response_model.dart';
+import 'package:business_banking/features/stocks_detail/api/stocks_service.dart';
+import 'package:business_banking/features/stocks_detail/api/stocks_service_response_model.dart';
 import 'package:clean_framework/clean_framework_defaults.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -17,8 +17,8 @@ void main() {
             "company": "Amazon.com",
             "symbol": "AMZN",
             "at_close": 89.38,
-            "shares" : 100,
-            "value" : 8938.0
+            "shares": 100,
+            "value": 8938.0
           }
         ]
       }),
