+++ business_banking/test/features/stocks/bloc/stocks_bloc_test.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
-import 'package:business_banking/features/stocks/bloc/stocks_bloc.dart';
-import 'package:business_banking/features/stocks/model/stocks_view_model.dart';
+import 'package:business_banking/features/stocks_detail/bloc/stocks_bloc.dart';
+import 'package:business_banking/features/stocks_detail/model/stocks_view_model.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
