+++ business_banking/test/features/stocks/model/stocks_entity_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:business_banking/features/stocks/model/stocks_entity.dart';
+import 'package:business_banking/features/stocks_detail/model/stocks_entity.dart';
 
 void main() {
   StocksEntity entity = StocksEntity(stocks: [
